SnapdragonCamera: Only set default video quality for rear camera
Currently, the default video quality is configured to be 4k UHD, but the
front-facing camera doesn't support this resolution, so the app crashes
when switching to the front-facing camera while in camcorder mode.

Set the specified default video quality only for the rear camera as a
quick workaround for the crash.